web-scraping/pyproject.toml
[project]
name = "web-scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "anthropic>=0.68.1",
    "beautifulsoup4>=4.14.0",
    "chromadb>=1.0.15",
    "fake-useragent>=2.2.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.30.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "nltk>=3.9.1",
    "numpy>=2.3.2",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "scrapy>=2.13.3",
    "scrapy-fake-useragent>=1.4.4",
    "selenium>=4.35.0",
    "spacy>=3.8.7",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
]
